+++ src/components/account/profile/index.jsx
@@ -1,78 +1,39 @@
-import { useFormContext } from 'react-hook-form';
-import { useEffect, useState } from 'react';
 import AccountProfileView from './AccountProfileView';
-import Users from '@/api/users';
-import useUserStore from '@/stores/useUserStore';
-import useModalState from '@/hooks/useModalState';
 import SuccessModal from '@/components/common/SuccessModal';
 import InitializeConfirmModal from '@/components/common/ConfirmModal';
+import useProfileModalLogic from '@/hooks/logic/useProfileChangeLogic';
 
-const AccountProfileLogic = () => {
-  const { modalState, toggleModal } = useModalState([
-    'profileChangeSuccessModal',
-    'initialConfirmModal',
-    'profileInitializeSuccessModal',
-  ]);
-  const { user } = useUserStore();
-  const { handleSubmit, formState, watch, setValue } = useFormContext();
-  const [isChanged, setIsChanged] = useState(false);
-  const [profileImageInit, setProfileImageInit] = useState(false);
-  const profileImageValue = watch('profileImageUrl');
-  const nicknameValue = watch('nickname');
-
-  const onProfileSubmit = async (data) => {
-    data.profileImageUrl =
-      data.profileImageUrl?.length === 0 ? user.profileImageUrl : data.profileImageUrl;
-    const { status } = await Users.edit(data);
-    if (status === 200) {
-      toggleModal('profileChangeSuccessModal');
-      setIsChanged(true);
-    }
-  };
-
-  const handleInitializeClick = async () => {
-    const initialImageData = {
-      profileImageUrl: null,
-      nickname: user.nickname,
-    };
-    const { status } = await Users.edit(initialImageData);
-    if (status === 200) {
-      setIsChanged(true);
-      setProfileImageInit(true);
-      setValue('profileImageUrl', null);
-      toggleModal('profileInitializeSuccessModal');
-    }
-  };
-
-  useEffect(() => {
-    if (user) {
-      if (
-        nicknameValue === user.nickname &&
-        (profileImageValue === user.profileImageUrl || profileImageValue?.length === 0)
-      ) {
-        setIsChanged(true);
-        return;
-      }
-      setIsChanged(false);
-    }
-  }, [nicknameValue, profileImageValue, user]);
-
+const AccountProfileChange = () => {
+  const {
+    user,
+    handleSubmit,
+    formState,
+    isChanged,
+    profileImageInit,
+    setProfileImageInit,
+    onProfileSubmit,
+    handleInitializeClick,
+    modalState,
+    toggleProfileChangeSuccessModal,
+    toggleProfileInitializeSuccessModal,
+    toggleInitialConfirmModal,
+  } = useProfileModalLogic();
   return (
     <>
       <SuccessModal
         desc='Profile modify success'
         isModalOpen={modalState.profileChangeSuccessModal}
-        closeModal={() => toggleModal('profileChangeSuccessModal')}
+        closeModal={toggleProfileChangeSuccessModal}
       />
       <SuccessModal
         desc='Profile image initialized'
         isModalOpen={modalState.profileInitializeSuccessModal}
-        closeModal={() => toggleModal('profileInitializeSuccessModal')}
+        closeModal={toggleProfileInitializeSuccessModal}
       />
       <InitializeConfirmModal
         desc='Are you sure you want to reset?'
         isModalOpen={modalState.initialConfirmModal}
-        closeModal={() => toggleModal('initialConfirmModal')}
+        closeModal={toggleInitialConfirmModal}
         handleInitializeClick={handleInitializeClick}
       />
       <AccountProfileView
@@ -83,10 +44,10 @@ const AccountProfileLogic = () => {
         handleInitializeClick={handleInitializeClick}
         profileImageInit={profileImageInit}
         setProfileImageInit={setProfileImageInit}
-        toggleModal={() => toggleModal('initialConfirmModal')}
+        toggleModal={toggleInitialConfirmModal}
       />
     </>
   );
 };
 
-export default AccountProfileLogic;
+export default AccountProfileChange;
